Rename js directory to ts
Current project uses typescript.

Issue #6 Create Blog Post Header(Title) component
